Refactor monolithic encodings/encodings.h
