+++ src/parquet/encodings/CMakeLists.txt
@@ -17,7 +17,8 @@
 
 # Headers: encodings
 install(FILES
-  encodings.h
+  decoder.h
+  encoder.h
   delta-bit-pack-encoding.h
   delta-byte-array-encoding.h
   delta-length-byte-array-encoding.h
